AppLink解析系の実装をAppLinkInfo側に移動
